Whoops, fix some issues with the `backend` rename: - Name the C++-side module binding correctly. - Fix leading-period `.mollytime` imports my find-all sweep missed.
